requirements.txt
fastapi
uvicorn
motor
pymongo
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
langchain
langgraph
scikit-learn
nltk
